[CT-002]: Move numeric_linear_chart to its own fragment
